+++ runlab/src/main/java/RunLab/Objects/Strava/Lap.java
@@ -1,0 +1,17 @@
+package RunLab.Objects.Strava;
+
+
+public class Lap {
+    private Long id;
+    private String name;
+    private float distance;
+    private float elapsed_time;
+    private float split;
+    private float max_speed;
+    private float average_speed;
+    private float max_heartrate;
+    private float average_heartrate;
+    private float pace_zone;
+
+    
+}
